리뷰 반영 https://github.com/yamanit/red-backend/pull/11
