+++ src/main/java/com/yamanit/red/user/repository/UserInMemoryRepository.java
@@ -1,0 +1,45 @@
+package com.yamanit.red.user.repository;
+
+import com.yamanit.red.user.model.User;
+import org.springframework.stereotype.Repository;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Repository
+public class UserInMemoryRepository implements UserRepository {
+	private static final Map<String, User> USER_MAP = new ConcurrentHashMap<>();
+
+	@Override
+	public Optional<User> findByUuid(String uuid) {
+		return Optional.ofNullable(USER_MAP.get(uuid));
+	}
+
+	@Override
+	public User addNewUser() {
+		String uuid = UUID.randomUUID().toString();
+
+		User newUser = new User(uuid);
+
+		USER_MAP.put(uuid, newUser);
+
+		return newUser;
+	}
+
+	@Override
+	public User getUser(String uuid) {
+		return USER_MAP.get(uuid);
+	}
+
+	@Override
+	public User save(User user) {
+		if (user.getUuid() == null) {
+			return addNewUser();
+		} else {
+			return USER_MAP.get(user.getUuid())
+						   .setNickname(user.getNickname());
+		}
+	}
+}
