--- src/main/java/com/yamanit/red/user/repository/UserRepositoryInterface.java
@@ -1,11 +1,0 @@
-package com.yamanit.red.user.repository;
-
-import com.yamanit.red.user.model.User;
-
-import java.util.Optional;
-
-public interface UserRepositoryInterface {
-	Optional<User> findByUuid(String uuid);
-
-	User addNewUser();
-}
